chenge in invoice

+++ src/Screens/Invoice/List/InvoiceListWrapper.tsx
@@ -1,3 +1,4 @@
+import Swal from 'sweetalert2';
 import { useDeleteInvoiceMutation, useGetAllInvoiceQuery } from '../../../Service/InvoiceApi/InvoiceApiSlice';
 import InvoiceList from './InvoiceList';
 
@@ -14,15 +15,28 @@ const InvoiceListWrapper = () => {
         return <div>Error fetching data</div>;
     }
 
-    const handleDelete = async (billId :string) => {
-        try {
-            await DeleteBill({ INVNo: billId });
-            console.log('Customer deleted successfully');
-            // Optionally, you can refetch the customer list or remove the deleted customer from the local state
-        } catch (error) {
-            console.error('Error deleting customer:', error);
-        }
-    };
+    
+    const handleDelete = async (billId: string) => {
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "Do you really want to delete this invoice?",
+          icon: 'question',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            try {
+                await DeleteBill({ INVNo: billId });
+              Swal.fire('Deleted!', 'The invoice has been deleted.', 'success');
+            } catch (error) {
+              Swal.fire('Error!', 'There was a problem deleting the invoice.', 'error');
+            }
+          }
+        });
+      };
+     
     return (
         <InvoiceList data={data} handleDelete={handleDelete} />
     )
